Category Wise
In this version, I have erased everything that I had done up till now and started over from scratch again because I had an idea of creating a separate page for every variable in the data. I have started from a category as it has few unique values and is it easy to handle.